python/grimlock/pyproject.toml
[tool.poetry]
name = "grimlock"
packages = [{include = "grimlock", from = ".."}]
version = "1.0.0"
description = "Versioned cryptographic operations module for privyy.io"
authors = ["Privyy <info@privyy.io>"]
readme = "README.md"
license = "ISC"
repository = "https://github.com/privyy-io/grimlock"
homepage = "https://github.com/privyy-io/grimlock"
keywords = ["cryptography", "encryption", "ecdh", "aes-gcm", "key-derivation", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: ISC License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]

[tool.poetry.dependencies]
python = "^3.12"
cryptography = "^42.0.0"
argon2-cffi = "^23.0.0"
mnemonic = "^0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
